Reorganize Reviews code
[#OSF-8576]
