+++ api/base/settings/defaults.py
@@ -93,7 +93,6 @@ INSTALLED_APPS = (
 
     # OSF
     'osf',
-    'reviews',
 
     # Addons
     'addons.osfstorage',
