+++ api/preprints/permissions.py
@@ -4,8 +4,8 @@ from rest_framework import exceptions
 
 from api.base.utils import get_user_auth
 from osf.models import PreprintService
+from osf.utils.workflows import DefaultStates
 from website.util import permissions as osf_permissions
-from reviews.workflow import States
 
 
 class PreprintPublishedOrAdmin(permissions.BasePermission):
@@ -21,7 +21,7 @@ class PreprintPublishedOrAdmin(permissions.BasePermission):
                 user_has_permissions = (obj.verified_publishable or
                     (node.is_public and auth.user.has_perm('view_submissions', obj.provider)) or
                     node.has_permission(auth.user, osf_permissions.ADMIN) or
-                    (node.is_contributor(auth.user) and obj.reviews_state != States.INITIAL.value)
+                    (node.is_contributor(auth.user) and obj.reviews_state != DefaultStates.INITIAL.value)
                 )
                 return user_has_permissions
         else:
