+++ api_tests/preprint_providers/views/test_preprint_provider_detail.py
@@ -1,11 +1,11 @@
 import pytest
 
 from api.base.settings.defaults import API_BASE
+from api.preprint_providers.permissions import GroupHelper
 from osf_tests.factories import (
     PreprintProviderFactory,
     AuthUserFactory,
 )
-from reviews.permissions import GroupHelper
 
 @pytest.mark.django_db
 class TestPreprintProviderExists:
