+++ api_tests/preprints/filters/test_filters.py
@@ -1,13 +1,11 @@
 import pytest
 
-from framework.auth.core import Auth
+from api.preprint_providers.permissions import GroupHelper
 from osf_tests.factories import (
     PreprintFactory,
     AuthUserFactory,
     SubjectFactory,
-    PreprintProviderFactory
 )
-from reviews.permissions import GroupHelper
 
 
 @pytest.mark.django_db
