+++ api_tests/preprints/views/test_preprint_actions.py
@@ -1,11 +1,10 @@
 import pytest
 
 from api.base.settings.defaults import API_BASE
+from api.preprint_providers.permissions import GroupHelper
 from osf_tests.factories import (
-    PreprintFactory,
     AuthUserFactory,
 )
-from reviews.permissions import GroupHelper
 from website.util import permissions as osf_permissions
 
 from api_tests.reviews.mixins.filter_mixins import ActionFilterMixin
