+++ api_tests/preprints/views/test_preprint_detail.py
@@ -1,12 +1,11 @@
-import functools
 import mock
-from django.db.models import Q
 import pytest
 
 from api.base.settings.defaults import API_BASE
 from api_tests import utils as test_utils
 from framework.auth.core import Auth
 from osf.models import PreprintService, NodeLicense
+from osf.utils.workflows import DefaultStates
 from osf_tests.factories import (
     PreprintFactory,
     AuthUserFactory,
@@ -15,10 +14,6 @@ from osf_tests.factories import (
     PreprintProviderFactory,
 )
 from rest_framework import exceptions
-from tests.base import fake, capture_signals
-from website.project.signals import contributor_added
-from website.identifiers.utils import build_ezid_metadata
-from reviews.workflow import States
 
 
 def build_preprint_update_payload(node_id, attributes=None, relationships=None):
@@ -69,7 +64,7 @@ class TestPreprintDetail:
 
     #   test contributors in preprint data
         assert data['relationships'].get('contributors', None)
-        assert data['relationships']['contributors'].get('data', None) == None
+        assert data['relationships']['contributors'].get('data', None) is None
 
     #   test node type and id in preprint data
         assert data['relationships']['node']['data'].get('id', None) == preprint.node._id
@@ -1014,15 +1009,15 @@ class TestReviewsPreprintDetailPermissions:
 
     @pytest.fixture()
     def unpublished_reviews_preprint(self, admin, reviews_provider, subject, public_project):
-        return PreprintFactory(creator=admin, filename='toe_socks_and_sunrises.pdf', provider=reviews_provider, subjects=[[subject._id]], project=public_project, is_published=False, reviews_state=States.PENDING.value)
+        return PreprintFactory(creator=admin, filename='toe_socks_and_sunrises.pdf', provider=reviews_provider, subjects=[[subject._id]], project=public_project, is_published=False, reviews_state=DefaultStates.PENDING.value)
 
     @pytest.fixture()
     def unpublished_reviews_initial_preprint(self, admin, reviews_provider, subject, public_project):
-        return PreprintFactory(creator=admin, filename='toe_socks_and_sunrises.pdf', provider=reviews_provider, subjects=[[subject._id]], project=public_project, is_published=False, reviews_state=States.INITIAL.value)
+        return PreprintFactory(creator=admin, filename='toe_socks_and_sunrises.pdf', provider=reviews_provider, subjects=[[subject._id]], project=public_project, is_published=False, reviews_state=DefaultStates.INITIAL.value)
 
     @pytest.fixture()
     def private_reviews_preprint(self, admin, reviews_provider, subject, private_project):
-        return PreprintFactory(creator=admin, filename='toe_socks_and_sunsets.pdf', provider=reviews_provider, subjects=[[subject._id]], project=private_project, is_published=False, reviews_state=States.PENDING.value)
+        return PreprintFactory(creator=admin, filename='toe_socks_and_sunsets.pdf', provider=reviews_provider, subjects=[[subject._id]], project=private_project, is_published=False, reviews_state=DefaultStates.PENDING.value)
 
     @pytest.fixture()
     def unpublished_url(self, unpublished_reviews_preprint):
