+++ api_tests/preprints/views/test_preprint_list.py
@@ -14,6 +14,7 @@ from api_tests.preprints.views.test_preprint_list_mixin import (
 from api_tests.reviews.mixins.filter_mixins import ReviewableFilterMixin
 from framework.auth.core import Auth
 from osf.models import PreprintService, Node
+from osf.utils.workflows import DefaultStates
 from osf_tests.factories import (
     ProjectFactory,
     PreprintFactory,
@@ -24,7 +25,6 @@ from osf_tests.factories import (
 from tests.base import ApiTestCase, capture_signals
 from website.project import signals as project_signals
 from website.util import permissions
-from reviews.workflow import States
 
 def build_preprint_create_payload(node_id=None, provider_id=None, file_id=None, attrs={}):
     payload = {
@@ -505,7 +505,7 @@ class TestReviewsPendingPreprintIsPublishedList(PreprintIsPublishedListMixin):
 
     @pytest.fixture()
     def preprint_unpublished(self, user_admin_contrib, provider_one, project_public, subject):
-        return PreprintFactory(creator=user_admin_contrib, filename='mgla.pdf', provider=provider_one, subjects=[[subject._id]], project=project_public, is_published=False, reviews_state=States.PENDING.value)
+        return PreprintFactory(creator=user_admin_contrib, filename='mgla.pdf', provider=provider_one, subjects=[[subject._id]], project=project_public, is_published=False, reviews_state=DefaultStates.PENDING.value)
 
     def test_unpublished_visible_to_admins(self, app, user_admin_contrib, preprint_unpublished, preprint_published, url):
         res = app.get(url, auth=user_admin_contrib.auth)
@@ -552,7 +552,7 @@ class TestReviewsInitialPreprintIsPublishedList(PreprintIsPublishedListMixin):
 
     @pytest.fixture()
     def preprint_unpublished(self, user_admin_contrib, provider_one, project_public, subject):
-        return PreprintFactory(creator=user_admin_contrib, filename='mgla.pdf', provider=provider_one, subjects=[[subject._id]], project=project_public, is_published=False, reviews_state=States.INITIAL.value)
+        return PreprintFactory(creator=user_admin_contrib, filename='mgla.pdf', provider=provider_one, subjects=[[subject._id]], project=project_public, is_published=False, reviews_state=DefaultStates.INITIAL.value)
 
     def test_unpublished_visible_to_admins(self, app, user_admin_contrib, preprint_unpublished, preprint_published, url):
         res = app.get(url, auth=user_admin_contrib.auth)
@@ -648,7 +648,7 @@ class TestReviewsInitialPreprintIsPublishedListMatchesDetail(PreprintListMatches
 
     @pytest.fixture()
     def preprint_unpublished(self, user_admin_contrib, provider_one, project_public, subject):
-        return PreprintFactory(creator=user_admin_contrib, filename='mgla.pdf', provider=provider_one, subjects=[[subject._id]], project=project_public, is_published=False, reviews_state=States.INITIAL.value)
+        return PreprintFactory(creator=user_admin_contrib, filename='mgla.pdf', provider=provider_one, subjects=[[subject._id]], project=project_public, is_published=False, reviews_state=DefaultStates.INITIAL.value)
 
     @pytest.fixture()
     def list_url(self):
@@ -701,7 +701,7 @@ class TestReviewsPendingPreprintIsPublishedListMatchesDetail(PreprintListMatches
 
     @pytest.fixture()
     def preprint_unpublished(self, user_admin_contrib, provider_one, project_public, subject):
-        return PreprintFactory(creator=user_admin_contrib, filename='mgla.pdf', provider=provider_one, subjects=[[subject._id]], project=project_public, is_published=False, reviews_state=States.PENDING.value)
+        return PreprintFactory(creator=user_admin_contrib, filename='mgla.pdf', provider=provider_one, subjects=[[subject._id]], project=project_public, is_published=False, reviews_state=DefaultStates.PENDING.value)
 
     @pytest.fixture()
     def list_url(self):
