+++ api_tests/reviews/mixins/comment_settings.py
@@ -1,15 +1,12 @@
-from datetime import timedelta
-
 import pytest
-from furl import furl
 
+from api.preprint_providers.permissions import GroupHelper
 from osf_tests.factories import (
     ActionFactory,
     AuthUserFactory,
     PreprintFactory,
     PreprintProviderFactory,
 )
-from reviews.permissions import GroupHelper
 from website.util import permissions as osf_permissions
