+++ osf/exceptions.py
@@ -114,3 +114,10 @@ class ValidationTypeError(ValidationError, MODMValidationTypeError):
 
 class NaiveDatetimeException(Exception):
     pass
+
+class InvalidTriggerError(Exception):
+    def __init__(self, trigger, state, valid_triggers):
+        self.trigger = trigger
+        self.state = state
+        self.valid_triggers = valid_triggers
+        self.message = 'Cannot trigger "{}" from state "{}". Valid triggers: {}'.format(trigger, state, valid_triggers)
