+++ osf/management/commands/create_fake_preprint_actions.py
@@ -8,8 +8,8 @@ from faker import Faker
 
 from django.core.management.base import BaseCommand
 
-from reviews import workflow
 from osf.models import Action, PreprintService, OSFUser
+from osf.utils.workflows import DefaultStates, DefaultTriggers
 
 logger = logging.getLogger(__name__)
 
@@ -44,8 +44,8 @@ class Command(BaseCommand):
             user = OSFUser.objects.get(guids___id=user_guid)
 
         fake = Faker()
-        triggers = [a.value for a in workflow.Triggers]
-        states = [s.value for s in workflow.States]
+        triggers = [a.value for a in DefaultTriggers]
+        states = [s.value for s in DefaultStates]
         for preprint in PreprintService.objects.filter(actions__isnull=True):
             for i in range(num_actions):
                 action = Action(
