+++ osf/models/action.py
@@ -5,11 +5,9 @@ from django.db import models
 
 from include import IncludeManager
 
-from reviews.workflow import Triggers
-from reviews.workflow import States
-
 from osf.models.base import BaseModel, ObjectIDMixin
 from osf.utils.fields import NonNaiveDateTimeField
+from osf.utils.workflows import DefaultStates, DefaultTriggers
 
 
 class Action(ObjectIDMixin, BaseModel):
@@ -19,9 +17,9 @@ class Action(ObjectIDMixin, BaseModel):
     target = models.ForeignKey('PreprintService', related_name='actions', on_delete=models.CASCADE)
     creator = models.ForeignKey('OSFUser', related_name='+', on_delete=models.CASCADE)
 
-    trigger = models.CharField(max_length=31, choices=Triggers.choices())
-    from_state = models.CharField(max_length=31, choices=States.choices())
-    to_state = models.CharField(max_length=31, choices=States.choices())
+    trigger = models.CharField(max_length=31, choices=DefaultTriggers.choices())
+    from_state = models.CharField(max_length=31, choices=DefaultStates.choices())
+    to_state = models.CharField(max_length=31, choices=DefaultStates.choices())
 
     comment = models.TextField(blank=True)
