+++ osf/models/mixins.py
@@ -1,11 +1,18 @@
 import pytz
+
 from django.apps import apps
-from django.db import models
+from django.db import models, transaction
 from django.core.exceptions import ObjectDoesNotExist
+from include import IncludeQuerySet
+
+from api.preprint_providers.workflows import Workflows, PUBLIC_STATES
 from framework.analytics import increment_user_activity_counters
+from osf.exceptions import InvalidTriggerError
 from osf.models.node_relation import NodeRelation
 from osf.models.nodelog import NodeLog
 from osf.models.tag import Tag
+from osf.utils.machines import ReviewsMachine
+from osf.utils.workflows import DefaultStates, DefaultTriggers
 from website.exceptions import NodeStateError
 from website import settings
 
@@ -456,3 +463,105 @@ class CommentableMixin(object):
         """Return extra data to pass as `params` to `Node.add_log` when a new comment is
         created, edited, deleted or restored."""
         return {}
+
+
+class ReviewProviderMixin(models.Model):
+    """A reviewed/moderated collection of objects.
+    """
+
+    REVIEWABLE_RELATION_NAME = None
+
+    class Meta:
+        abstract = True
+
+    reviews_workflow = models.CharField(null=True, blank=True, max_length=15, choices=Workflows.choices())
+    reviews_comments_private = models.NullBooleanField()
+    reviews_comments_anonymous = models.NullBooleanField()
+
+    @property
+    def is_reviewed(self):
+        return self.reviews_workflow is not None
+
+    def get_reviewable_state_counts(self):
+        assert self.REVIEWABLE_RELATION_NAME, 'REVIEWABLE_RELATION_NAME must be set to compute state counts'
+        qs = getattr(self, self.REVIEWABLE_RELATION_NAME)
+        if isinstance(qs, IncludeQuerySet):
+            qs = qs.include(None)
+        qs = qs.filter(node__isnull=False, node__is_deleted=False, node__is_public=True).values('reviews_state').annotate(count=models.Count('*'))
+        counts = {state.value: 0 for state in DefaultStates}
+        counts.update({row['reviews_state']: row['count'] for row in qs if row['reviews_state'] in counts})
+        return counts
+
+    def add_admin(self, user):
+        from api.preprint_providers.permissions import GroupHelper
+        return GroupHelper(self).get_group('admin').user_set.add(user)
+
+    def add_moderator(self, user):
+        from api.preprint_providers.permissions import GroupHelper
+        return GroupHelper(self).get_group('moderator').user_set.add(user)
+
+
+class ReviewableMixin(models.Model):
+    """Something that may be included in a reviewed collection and is subject to a reviews workflow.
+    """
+
+    class Meta:
+        abstract = True
+
+    # NOTE: reviews_state should rarely/never be modified directly -- use the state transition methods below
+    reviews_state = models.CharField(max_length=15, db_index=True, choices=DefaultStates.choices(), default=DefaultStates.INITIAL.value)
+
+    date_last_transitioned = models.DateTimeField(null=True, blank=True, db_index=True)
+
+    @property
+    def in_public_reviews_state(self):
+        public_states = PUBLIC_STATES.get(self.provider.reviews_workflow)
+        if not public_states:
+            return False
+        return self.reviews_state in public_states
+
+    def reviews_submit(self, user):
+        """Run the 'submit' state transition and create a corresponding Action.
+
+        Params:
+            user: The user triggering this transition.
+        """
+        return self.__run_transition(DefaultTriggers.SUBMIT.value, user=user)
+
+    def reviews_accept(self, user, comment):
+        """Run the 'accept' state transition and create a corresponding Action.
+
+        Params:
+            user: The user triggering this transition.
+            comment: Text describing why.
+        """
+        return self.__run_transition(DefaultTriggers.ACCEPT.value, user=user, comment=comment)
+
+    def reviews_reject(self, user, comment):
+        """Run the 'reject' state transition and create a corresponding Action.
+
+        Params:
+            user: The user triggering this transition.
+            comment: Text describing why.
+        """
+        return self.__run_transition(DefaultTriggers.REJECT.value, user=user, comment=comment)
+
+    def reviews_edit_comment(self, user, comment):
+        """Run the 'edit_comment' state transition and create a corresponding Action.
+
+        Params:
+            user: The user triggering this transition.
+            comment: New comment text.
+        """
+        return self.__run_transition(DefaultTriggers.EDIT_COMMENT.value, user=user, comment=comment)
+
+    def __run_transition(self, trigger, **kwargs):
+        reviews_machine = ReviewsMachine(self, 'reviews_state')
+        trigger_fn = getattr(reviews_machine, trigger)
+        with transaction.atomic():
+            result = trigger_fn(**kwargs)
+            action = reviews_machine.action
+            if not result or action is None:
+                valid_triggers = reviews_machine.get_triggers(self.reviews_state)
+                raise InvalidTriggerError(trigger, self.reviews_state, valid_triggers)
+            return action
