+++ osf/models/preprint_provider.py
@@ -4,15 +4,13 @@ from django.db import models
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
+from api.preprint_providers.permissions import GroupHelper, PERMISSIONS
 from osf.models.base import BaseModel, ObjectIDMixin
 from osf.models.licenses import NodeLicense
+from osf.models.mixins import ReviewProviderMixin
 from osf.models.subject import Subject
 from osf.utils.datetime_aware_jsonfield import DateTimeAwareJSONField
 from osf.utils.fields import EncryptedTextField
-
-from reviews import permissions as reviews_permissions
-from reviews.models import ReviewProviderMixin
-
 from website import settings
 from website.util import api_v2_url
 
@@ -62,7 +60,7 @@ class PreprintProvider(ObjectIDMixin, ReviewProviderMixin, BaseModel):
                                         null=True, blank=True, on_delete=models.CASCADE)
 
     class Meta:
-        permissions = tuple(reviews_permissions.PERMISSIONS.items()) + (
+        permissions = tuple(PERMISSIONS.items()) + (
             # custom permissions for use in the OSF Admin App
             ('view_preprintprovider', 'Can view preprint provider details'),
         )
@@ -129,4 +127,4 @@ def rules_to_subjects(rules):
 @receiver(post_save, sender=PreprintProvider)
 def create_provider_auth_groups(sender, instance, created, **kwargs):
     if created:
-        reviews_permissions.GroupHelper(instance).update_provider_auth_groups()
+        GroupHelper(instance).update_provider_auth_groups()
