--- reviews/apps.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
-from django.apps import AppConfig
-
-
-class ReviewsConfig(AppConfig):
-    name = 'reviews'
