--- reviews/models/__init__.py
@@ -1,2 +1,0 @@
-from .mixins import ReviewProviderMixin  # noqa
-from .mixins import ReviewableMixin  # noqa
